🚚 설계에 맞게 폴더구조 변경
* issue: [NAILCASE-113](https://nailcase.atlassian.net/browse/NAILCASE-113)
sub-issues:
- [NAILCASE-117](https://nailcase.atlassian.net/browse/NAILCASE-117)
